+++ shop/models/defaults/bases.py
@@ -190,16 +190,20 @@ class BaseCart(models.Model):
         self.extra_price_fields = [] # Reset the price fields
         self.subtotal_price = Decimal('0.0') # Reset the subtotal
 
+        # This will hold extra information that cart modifiers might want to pass
+        # to each other
+        state = {} 
+
         for item in items: # For each OrderItem (order line)...
             item.product = products_dict[item.product_id] #This is still the ghetto select_related
-            self.subtotal_price = self.subtotal_price + item.update(self)
+            self.subtotal_price = self.subtotal_price + item.update(state)
             item.save()
         
         self.current_total = self.subtotal_price
         # Now we have to iterate over the registered modifiers again (unfortunately)
         # to pass them the whole Order this time
         for modifier in cart_modifiers_pool.get_modifiers_list():
-            modifier.process_cart(self)
+            modifier.process_cart(self, state)
         
         self.total_price = self.current_total
 
@@ -246,7 +250,7 @@ class BaseCartItem(models.Model):
         self.line_total = Decimal('0.0')
         self.current_total = Decimal('0.0') # Used by cart modifiers
 
-    def update(self, cart):
+    def update(self, state):
         self.extra_price_fields = [] # Reset the price fields
         self.line_subtotal = self.product.get_price() * self.quantity
         self.current_total = self.line_subtotal
@@ -254,7 +258,7 @@ class BaseCartItem(models.Model):
         for modifier in cart_modifiers_pool.get_modifiers_list():
             # We now loop over every registered price modifier,
             # most of them will simply add a field to extra_payment_fields
-            modifier.process_cart_item(self, cart)
+            modifier.process_cart_item(self, state)
         
         self.line_total = self.current_total
         return self.line_total